% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{customer_data}
\alias{customer_data}
\title{Customer Data}
\format{
## `customer_data`
A list of 6 objects:
\describe{
  \item{Fresh}{annual spending (m.u.) on fresh products (Continuous);}
  \item{Milk}{annual spending (m.u.) on milk products (Continuous);}
  \item{Grocery}{annual spending (m.u.)on grocery products (Continuous);}
  \item{Frozen}{annual spending (m.u.)on frozen products (Continuous);}
  \item{Detergents_Paper}{annual spending (m.u.) on detergents and paper products (Continuous);}
  \item{Delicassen}{annual spending (m.u.)on and delicatessen products (Continuous).}
}
}
\source{
<www.archive.ics.uci.edu/dataset/292/wholesale+customers/>
}
\usage{
customer_data
}
\description{
The data set refers to clients of a wholesale distributor. 
It includes the annual spending in monetary units (m.u.) on diverse product categories.
}
\keyword{datasets}
