% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DESPOTA.R
\name{despota}
\alias{despota}
\title{despota}
\usage{
despota(
  data,
  distmat = NULL,
  distMethod = "euclidean",
  agglMethod = "ward.D2",
  M = 999,
  alpha = 0.01,
  listVal = FALSE,
  seed = NULL,
  par = TRUE,
  ncores = NULL
)
}
\arguments{
\item{data}{a matrix or data.frame of data}

\item{distmat}{a dissimilarity matrix obtained through the `dist()` function}

\item{distMethod}{distance method to use, one of "euclidean", "maximum", 
"manhattan", "canberra", "binary" or "minkowski" (default "euclidean")}

\item{agglMethod}{agglomerative method to use, one of "ward.D", "ward.D2", 
"single", "complete", "average", "mcquitty", "median"or "centroid"
 (default "ward.D2")}

\item{M}{number of permutations (default 999)}

\item{alpha}{significant level associated to the testing procedure
(default 0.01)}

\item{listVal}{logical to add additional information regarding the splitting 
procedure (default FALSE)}

\item{seed}{seed for the permutation}

\item{par}{logical to perform parallel computing (default TRUE)}

\item{ncores}{number of cores to be used. Default `ncores = NULL` automatically 
chooses the number of all possible physical cores minus one.}
}
\value{
A list of two elements, `Arguments` and `DF`. The first contains the 
dendrogram branches and splits in a dendextend format and additional information 
of the method used; the second object is a data.frame containing the observations, 
the cluster labels and their related p-values.
When `listVal = TRUE`, the output has an additional object containing the info 
regarding test statistics in each split.
}
\description{
DESPOTA hierarchical clustering of a dataset or a dissimilarity matrix
}
\details{
Frontend function that fits a dendrogram and finds the optimal cut through the DESPOTA procedure.
}
\examples{
data <- c(1,2,3,4,5, 7,9,10,11,12,  19,24,28,32,38, 54)
despota(data, distMethod = "euclidean", agglMethod = "ward.D2", 
M = 999, alpha = 0.05, seed = 1234, par = FALSE)

}
