% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DESPOTApar.R
\name{despotaPAR3}
\alias{despotaPAR3}
\title{DespotaPar}
\usage{
despotaPAR3(
  data,
  distmat = NULL,
  distMethod = "euclidean",
  agglMethod = "ward.D2",
  M = 999,
  alpha = 0.01,
  listVal = FALSE,
  seed = NULL,
  ...
)
}
\arguments{
\item{data}{}

\item{distmat}{}

\item{distMethod}{}

\item{agglMethod}{}

\item{M}{}

\item{alpha}{}

\item{listVal}{}

\item{seed}{}

\item{...}{}
}
\value{
gatto
}
\description{
DespotaPar
}
\examples{
#cane


}
