% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{findpos}
\alias{findpos}
\title{despotaPar}
\usage{
findpos(i, matrixNodes)
}
\arguments{
\item{i}{number of rows to be matched}

\item{matrixNodes}{data matrix with the necessary data}
}
\description{
Despota parallelized
}
\keyword{internal}
