% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plot.despota}
\alias{plot.despota}
\title{plot.despota}
\usage{
\method{plot}{despota}(x, main = NULL, sub = NULL, ...)
}
\arguments{
\item{x}{a despota object}

\item{main}{plot title}

\item{sub}{plot subtitle}

\item{...}{additional options passed to the plot function}
}
\value{
a dendrogram of the despota model
}
\description{
Plot a despota class object
}
