% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{customer_data}
\alias{customer_data}
\title{Customer Data}
\format{
## `customer_data`
A list of 6 objects:
\describe{
  \item{Fresh}{...}
  \item{Milk}{...}
  \item{Grocery}{...}
  \item{Frozen}{...}
  \item{Detergents_Paper}{...}
  \item{Delicassen}{...}
}
}
\source{
<www.google.it>
}
\usage{
customer_data
}
\description{
An example dataset ....put the description HERE....
}
\keyword{datasets}
